content/html/content/test/test_bug546995-5.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=546995
-->
<head>
  <title>Test for Bug 546995</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=546995">Mozilla Bug 546995</a>
<p id="display"></p>
<script type="application/javascript">
  document.body.focus();
  is(document.activeElement, document.body,
     "The document body should have the focus");
</script>
<div id='content'>
  <input id='i' autofocus>
</div>
<pre id="test">
<script type="application/javascript;version=1.8">

/** Test for Bug 546995 **/

SimpleTest.waitForExplicitFinish();

document.getElementById('i').addEventListener("focus", function(aEvent) {
  aEvent.target.removeEventListener("focus", arguments.callee, false);

  ok(true, "The input element has been focused");
  SimpleTest.finish();
  }, false);

</script>
</pre>
</body>
</html>
